connect_backend/requirements.txt
charset-normalizer==3.3.0
demjson3==3.0.6
Django==5.0.1
djangorestframework==3.14.0
django-prometheus==2.3.1
json5==0.9.14
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.17.3
junit-xml==1.9
mysqlclient==2.2.1
pbr==5.11.1
pyrsistent==0.19.3
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.8.8
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
virtualenv==20.24.5
yarl==1.9.2
requests==2.31.0
